- Исправления 1.16.1 - Причесал проект по MVVM, стало аккуратнее.
