+++ Tracker/ViewControllers/CreatingNewOneOffTrackerVC/Views/CreatingOneOffTrackerVC+TextFieldDelegate.swift
@@ -1,0 +1,31 @@
+//
+//  CreatingOneOffTrackerVC+TextFieldDelegate.swift
+//  Tracker
+//
+//  Created by Kirill Sklyarov on 08.04.2024.
+//
+
+import UIKit
+
+// MARK: - TextFieldDelegate - control of TextField length
+extension CreatingOneOffTrackerVC: UITextFieldDelegate {
+    
+    func textField(_ textField: UITextField, shouldChangeCharactersIn range: NSRange, replacementString string: String) -> Bool {
+        let currentCharacterCount = textField.text?.count ?? 0
+        if currentCharacterCount <= 38 {
+            hideLabelExceedTextFieldLimit()
+            isCreateButtonEnable()
+            textField.textColor = .black
+            return true
+        } else {
+            print("Check: opps")
+            showLabelExceedTextFieldLimit()
+            textField.textColor = .red
+            return true
+        }
+    }
+    
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        trackerNameTextField.resignFirstResponder()
+    }
+}
