+++ Tracker/ViewControllers/CreatingNewTrackerVC/CreatingNewTrackerViewController.swift
@@ -257,7 +257,7 @@ extension CreatingNewTrackerViewController: UITableViewDataSource, UITableViewDe
         } else {
             let scheduleVC = ScheduleViewController()
             let navVC = UINavigationController(rootViewController: scheduleVC)
-            scheduleVC.scheduleToPass = { [weak self] schedule in
+            scheduleVC.viewModel.scheduleToPass = { [weak self] schedule in
                 guard let self = self,
                       let cell = tableView.cellForRow(at: indexPath) else { return }
                 cell.detailTextLabel?.text = schedule
