+++ Tracker/ViewControllers/EditingCategoryVC/Views/EditingCategoryVC+TextField.swift
@@ -1,0 +1,15 @@
+//
+//  EditingCategoryVC+TexField.swift
+//  Tracker
+//
+//  Created by Kirill Sklyarov on 08.04.2024.
+//
+
+import UIKit
+
+// MARK: - UITextFieldDelegate
+extension EditingCategoryViewController: UITextFieldDelegate {
+    func textFieldShouldReturn(_ textField: UITextField) -> Bool {
+        categoryNameTextField.resignFirstResponder()
+    }
+}
