--- Tracker/ViewControllers/ScheduleVC/ScheduleViewController.swift
@@ -1,137 +1,0 @@
-//
-//  ScheduleViewController.swift
-//  Tracker
-//
-//  Created by Kirill Sklyarov on 14.03.2024.
-//
-
-import UIKit
-
-final class ScheduleViewController: UIViewController {
-    
-    // MARK: - UI Properties
-    private let tableView = UITableView()
-    private lazy var doneButton = setupButtons(title: "Готово")
-    
-    private let weekdays = ["Понедельник", "Вторник", "Среда", "Четверг", "Пятница", "Суббота", "Воскресенье"]
-    
-    // MARK: - Private Properties
-    private let rowHeight = CGFloat(75)
-    private var arrayOfIndexes = [Int]()
-    
-    var scheduleToPass: ( (String) -> Void )?
-    
-    // MARK: - Life Cycles
-    override func viewDidLoad() {
-        super.viewDidLoad()
-        
-        setupUI()
-        
-        setupTableView()
-        
-    }
-    
-    // MARK: - IB Actions
-    @objc private func weekDaySwitchValueChanged(_ sender: UISwitch) {
-        guard let cell = sender.superview as? UITableViewCell,
-              let indexPath = tableView.indexPath(for: cell) else { return }
-        if sender.isOn {
-            arrayOfIndexes.append(indexPath.row)
-            print(arrayOfIndexes)
-        } else {
-            arrayOfIndexes.removeAll(where: { $0 == indexPath.row })
-            print(arrayOfIndexes)
-        }
-    }
-    
-    @objc private func scheduleDoneButtonTapped(_ sender: UIButton) {
-        let scheduleString = getStringFromArray(array: arrayOfIndexes)
-        scheduleToPass?(scheduleString)
-        dismiss(animated: true)
-    }
-    
-    // MARK: - Private Methods
-    private func setupUI() {
-        
-        let tableViewHeight = CGFloat(weekdays.count) * rowHeight
-        
-        self.title = "Расписание"
-        view.backgroundColor = UIColor(named: "projectBackground")
-        
-        view.addSubViews([tableView, doneButton])
-        
-        NSLayoutConstraint.activate([
-            tableView.topAnchor.constraint(equalTo: view.safeAreaLayoutGuide.topAnchor, constant: 30),
-            tableView.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            tableView.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
-            tableView.heightAnchor.constraint(equalToConstant: tableViewHeight),
-            
-            doneButton.leadingAnchor.constraint(equalTo: view.leadingAnchor, constant: 16),
-            doneButton.trailingAnchor.constraint(equalTo: view.trailingAnchor, constant: -16),
-            doneButton.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -16),
-        ])
-    }
-    
-    private func setupTableView() {
-        tableView.dataSource = self
-        tableView.delegate = self
-        tableView.register(UITableViewCell.self, forCellReuseIdentifier: "cell")
-        
-        tableView.layer.cornerRadius = 16
-        tableView.separatorInset = UIEdgeInsets(top: 0, left: 16, bottom: 0, right: 16)
-        tableView.tableHeaderView = UIView()
-    }
-    
-    private func setupButtons(title: String) -> UIButton {
-        let button = UIButton()
-        button.setTitle(title, for: .normal)
-        button.titleLabel?.font = .systemFont(ofSize: 16, weight: .medium)
-        button.setTitleColor(.white, for: .normal)
-        button.backgroundColor = .black
-        button.layer.masksToBounds = true
-        button.layer.cornerRadius = 15
-        button.translatesAutoresizingMaskIntoConstraints = false
-        button.heightAnchor.constraint(equalToConstant: 60).isActive = true
-        button.addTarget(self, action: #selector(scheduleDoneButtonTapped), for: .touchUpInside)
-        return button
-    }
-    
-    private func getStringFromArray(array: [Int]) -> String {
-        if array.count == 7 { return "Каждый день"} else {
-            let daysOfWeek = ["Пн", "Вт", "Ср", "Чт", "Пт", "Сб", "Вс"]
-            let arrayOfStrings = array.map { daysOfWeek[$0] }
-            let resultString = arrayOfStrings.joined(separator: ", ")
-            return resultString
-        }
-    }
-    
-}
-
-// MARK: - UITableViewDataSource, UITableViewDelegate
-extension ScheduleViewController: UITableViewDataSource, UITableViewDelegate {
-    
-    func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        weekdays.count
-    }
-    
-    func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        let cell = tableView.dequeueReusableCell(withIdentifier: "cell", for: indexPath)
-        cell.selectionStyle = .none
-        let weekDayswitch = UISwitch()
-        weekDayswitch.onTintColor = UIColor(named: "switchOnColor")
-        weekDayswitch.addTarget(self, action: #selector(weekDaySwitchValueChanged), for: .valueChanged)
-        cell.accessoryView = weekDayswitch
-        cell.textLabel?.text = weekdays[indexPath.row]
-        cell.backgroundColor = UIColor(named: "textFieldBackgroundColor")
-        
-        if indexPath.row == weekdays.count - 1 {
-            cell.separatorInset = UIEdgeInsets(top: 0, left: 0, bottom: 0, right: .greatestFiniteMagnitude)
-        }
-        
-        return cell
-    }
-    
-    func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        rowHeight
-    }
-}
