+++ Tracker/ViewControllers/TabBarVC/TabBarController.swift
@@ -9,13 +9,11 @@ import UIKit
 
 final class TabBarController: UITabBarController {
     
-    
     override func viewDidLoad() {
         super.viewDidLoad()
         
         setupTabBarController()
         
-       
     }
     
     func setupTabBarController() {
@@ -25,7 +23,7 @@ final class TabBarController: UITabBarController {
         
         trackerVC.tabBarItem = UITabBarItem(title: "Трекер", image: UIImage(systemName: "record.circle.fill"), tag: 0)
         statisticVC.tabBarItem = UITabBarItem(title: "Статистика", image: UIImage(systemName: "hare.fill"), tag: 1)
-                
+        
         self.viewControllers = [trackerVC, statisticVC]
         
         self.tabBar.backgroundColor = .white
